Move path validation to resources.py
